refactor: move db diagnostic scripts into package
