+++ db_tools/script_database_validator.py
