+++ db_tools/temp_db_check.py
