+++ scripts/final_root_assessment.py
@@ -36,8 +36,6 @@ def final_root_assessment():
         "ENTERPRISE_DEVELOPMENT_MISSION_ACCOMPLISHED.md": "Mission status",
         "FINAL_LESSONS_LEARNED_INTEGRATION_VALIDATION.md": "Lessons learned",
         "UNIFIED_WRAP_UP_ORCHESTRATOR_IMPLEMENTATION_COMPLETE.md": "Wrap-up status",
-        # Temporary validation files (acceptable)
-        "temp_db_check.py": "Temporary database validation script",
     }
 
     # Scan actual root files
